// src/app/auth/models/UserModel.ts
export class UserModel {
    id?: string;
    email?: string;
    firstName?: string;
    lastName?: string;
    phoneNumber?: string;
    name?: string;
    imageUrl?: string | null;
    status?: string;
    vendorId?: string;
    emailConfirmed?: boolean;
    roles?: string[];
    wengazPercentage?:string;
    wengazPoint?:string;
    Provider?:string;
  }
  